+++ R/map_layer_path.R
@@ -20,15 +20,19 @@ mapdeckPathDependency <- function() {
 #' @param stroke_width width of the stroke in meters. Default 1.
 #' @param dash_size size of each dash, relative to the width of the stroke
 #' @param dash_gap size of the gap between dashes, relative to the width of the stroke
-#' @param offest The offset to draw each path with, relative to the width of the path.
+#' @param offset The offset to draw each path with, relative to the width of the path.
 #' Negative offset is to the left hand side, and positive offset is to the right hand side.
 #' 0 extrudes the path so that it is centered at the specified coordinates.
 #' @param billboard logical indicating if the path always faces the camera (TRUE) or
 #' if it always faces up (FALSE)
-#' @param width_units
-#' @param width_scale
-#' @param width_min_pixels
-#' @param width_max_pixels
+#' @param width_units The units of the line width, one of 'meters', 'pixels'.
+#' When zooming in and out, meter sizes scale with the base map, and pixel sizes
+#' remain the same on screen.
+#' @param width_scale The path width multiplier that multiplied to all paths.
+#' @param width_min_pixels The minimum path width in pixels.
+#' This can be used to prevent the path from getting too thin when zoomed out.
+#' @param width_max_pixels The maximum path width in pixels.
+#' his prop can be used to prevent the path from getting too thick when zoomed in.
 #'
 #' @inheritSection add_polygon data
 #' @inheritSection add_arc legend
@@ -140,6 +144,9 @@ add_path <- function(
 		l[["bbox"]] <- NULL
 	}
 
+	use_offset <- !is.null( offset )
+	use_dash <- !is.null( dash_size ) | !is.null( dash_gap )
+
 	layer_id <- layerId(layer_id, "path")
 	checkHexAlpha( highlight_colour )
 
@@ -168,7 +175,8 @@ add_path <- function(
 	invoke_method(
 		map, jsfunc, map_type( map ), shape[["data"]], layer_id, auto_highlight,
 		highlight_colour, shape[["legend"]], bbox, update_view, focus_layer,
-		js_transitions, billboard, brush_radius, width_units, width_scale, width_min_pixels, width_max_pixels
+		js_transitions, billboard, brush_radius, width_units, width_scale, width_min_pixels,
+		width_max_pixels, use_offset, use_dash
 		)
 }