+++ NEWS.md
@@ -1,7 +1,7 @@
 # v0.3.3
 
+* path width options [issue 300](https://github.com/SymbolixAU/mapdeck/issues/300)
 * min and max pitch and zoom options [issue 295](https://github.com/SymbolixAU/mapdeck/issues/295)
-
 
 # v0.3.2